pyproject.toml
[tool.poetry]
name = "sessionbot"
version = "0.1.0"
description = ""
authors = ["Slava Kuryaev <92830706+necrosskull@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "13.15"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
python-logging-loki = "^0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
